extract composition logic into `ToDoUIComposer`
